requirements.txt
ultralytics>=8.3.0
opencv-python-headless>=4.10.0.84
pillow>=10.0.0
numpy>=1.26.0
